refactor: change the order of importing/executing solara script
We now do this after solara.server.starlette is imported, making
the starlette app variable accessible from the solara script or module.
This makes it easier to add custom endpoints to a solara app.
