+++ solara/server/app.py
@@ -62,6 +62,35 @@ class AppScript:
         else:
             self.name = name
         self.path: Path = Path(self.name).resolve()
+        if self.path.is_dir():
+            self.type = AppType.DIRECTORY
+            # resolve the directory, because Path("file").parent.parent == "." != ".."
+            self.directory = self.path.resolve()
+        elif self.name.endswith(".py"):
+            self.type = AppType.SCRIPT
+            self.directory = self.path.parent.resolve()
+        elif self.name.endswith(".ipynb"):
+            self.type = AppType.NOTEBOOK
+            self.directory = self.path.parent.resolve()
+        else:
+            self.type = AppType.MODULE
+            try:
+                spec = importlib.util.find_spec(self.name)
+            except ValueError:
+                if self.name not in sys.modules:
+                    raise ImportError(f"Module {self.name} not found")
+                spec = importlib.util.spec_from_file_location(self.name, sys.modules[self.name].__file__)
+            if spec is None:
+                raise ImportError(f"Module {self.name} cannot be found")
+            assert spec is not None
+            if spec.origin is None:
+                raise ImportError(f"Module {self.name} cannot be found, or is a namespace package")
+            assert spec.origin is not None
+            self.path = Path(spec.origin)
+            self.directory = self.path.parent
+        self._initialized = False
+
+    def init(self):
         try:
             context = kernel_context.get_current_context()
         except RuntimeError:
@@ -84,6 +113,7 @@ class AppScript:
                 package_root_path = Path(mod.__file__).parent
                 reload.reloader.root_path = package_root_path
         dummy_kernel_context.close()
+        self._initialized = True
 
     def _execute(self):
         logger.info("Executing %s", self.name)
@@ -97,10 +127,8 @@ class AppScript:
                 if working_directory not in sys.path:
                     sys.path.insert(0, working_directory)
 
-        if self.path.is_dir():
-            self.type = AppType.DIRECTORY
+        if self.type == AppType.DIRECTORY:
             # resolve the directory, because Path("file").parent.parent == "." != ".."
-            self.directory = self.path.resolve()
             routes = solara.generate_routes_directory(self.path)
 
             if any(name for name in sys.modules.keys() if name.startswith(self.name)):
@@ -110,23 +138,19 @@ class AppScript:
                     "can avoid this ambiguity."
                 )
 
-        elif self.name.endswith(".py"):
-            self.type = AppType.SCRIPT
+        elif self.type == AppType.SCRIPT:
             add_path()
             # manually add the script to the watcher
             reload.reloader.watcher.add_file(self.path)
-            self.directory = self.path.parent.resolve()
             initial_namespace = {
                 "__name__": "__main__",
             }
             with reload.reloader.watch():
                 routes = [solara.autorouting._generate_route_path(self.path, first=True, initial_namespace=initial_namespace)]
-        elif self.name.endswith(".ipynb"):
-            self.type = AppType.NOTEBOOK
+        elif self.type == AppType.NOTEBOOK:
             add_path()
             # manually add the notebook to the watcher
             reload.reloader.watcher.add_file(self.path)
-            self.directory = self.path.parent.resolve()
             with reload.reloader.watch():
                 routes = [solara.autorouting._generate_route_path(self.path, first=True)]
         else:
@@ -134,21 +158,6 @@ class AppScript:
             # automatically
             with reload.reloader.watch():
                 self.type = AppType.MODULE
-                try:
-                    spec = importlib.util.find_spec(self.name)
-                except ValueError:
-                    if self.name not in sys.modules:
-                        raise ImportError(f"Module {self.name} not found")
-                    spec = importlib.util.spec_from_file_location(self.name, sys.modules[self.name].__file__)
-                if spec is None:
-                    raise ImportError(f"Module {self.name} cannot be found")
-                assert spec is not None
-                if spec.origin is None:
-                    raise ImportError(f"Module {self.name} cannot be found, or is a namespace package")
-                assert spec.origin is not None
-                self.path = Path(spec.origin)
-                self.directory = self.path.parent
-
                 mod = importlib.import_module(self.name)
                 routes = solara.generate_routes(mod)
 
@@ -213,7 +222,12 @@ class AppScript:
         for context in context_values:
             context.close()
 
+    def check(self):
+        if not self._initialized:
+            raise RuntimeError("Call solara.server.app.ensure_apps_initialized() first")
+
     def run(self):
+        self.check()
         if reload.reloader.requires_reload or self._first_execute_app is None:
             with thread_lock:
                 if reload.reloader.requires_reload or self._first_execute_app is None:
@@ -489,3 +503,9 @@ patch.patch()
 if "SOLARA_APP" in os.environ:
     with pdb_guard():
         apps["__default__"] = AppScript(os.environ.get("SOLARA_APP", "solara.website.pages:Page"))
+
+
+@solara.util.once
+def ensure_apps_initialized():
+    for app in apps.values():
+        app.init()
