+++ solara/server/flask.py
@@ -291,3 +291,7 @@ if has_solara_enterprise:
 
 if __name__ == "__main__":
     app.run(debug=False, port=8765)
+
+# we can only call this at the module level, which means that the solara script cannot import this
+# module. This is a difference with the asgi standard, which provides a lifecycle hook (see starlette.py)
+appmod.ensure_apps_initialized()
